move test runner to apex dir
